adding jet corrections

+++ bg/scripts/split_nlp_big_skims.sh
@@ -17,7 +17,12 @@ cmsenv
 export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:$CMSSW_BASE/src/cms-tools/lib/classes"
 
 
-for f in ST_t-channel_antitop.root ST_t-channel_top.root TTJets_SingleLeptFromT_TuneCUETP8M1.root TTJets_DiLept_TuneCUETP8M1.root TTJets_SingleLeptFromTbar_TuneCUETP8M1.root; do
-    echo $SCRIPTS_WD/skim_split.py -i $NLP_SKIM_OUTPUT_DIR/sum/type_sum/$f | tee /tmp/$f 
-    $SCRIPTS_WD/skim_split.py -i $NLP_SKIM_OUTPUT_DIR/sum/type_sum/$f | tee /tmp/$f 
+for f in ST_t-channel_antitop.root ST_t-channel_top.root TTJets_SingleLeptFromT_TuneCUETP8M1.root TTJets_DiLept_TuneCUETP8M1.root TTJets_SingleLeptFromTbar_TuneCUETP8M1.root WJetsToLNu_HT-1200To2500_TuneCUETP8M1.root WJetsToLNu_HT-200To400_TuneCUETP8M1.root WJetsToLNu_HT-400To600_TuneCUETP8M1.root WJetsToLNu_HT-600To800_TuneCUETP8M1.root WJetsToLNu_HT-800To1200_TuneCUETP8M1.root ZJetsToNuNu_HT-100To200_13TeV-madgraph.root ZJetsToNuNu_HT-200To400_13TeV-madgraph.root ZJetsToNuNu_HT-400To600_13TeV-madgraph.root ZJetsToNuNu_HT-600To800_13TeV-madgraph.root ZJetsToNuNu_HT-800To1200_13TeV-madgraph.root; do
+    if [ -f "$NLP_SKIM_OUTPUT_DIR/sum/type_sum/$f" ]; then
+        echo "$f exists."
+        echo $SCRIPTS_WD/skim_split.py -i $NLP_SKIM_OUTPUT_DIR/sum/type_sum/$f | tee /tmp/$f 
+        $SCRIPTS_WD/skim_split.py -i $NLP_SKIM_OUTPUT_DIR/sum/type_sum/$f | tee /tmp/$f 
+    else 
+        echo "$f does not exist. Skipping."
+    fi
 done
